unity: Only set responder's TS to 0.0.0.0/0 if we received such a TS
iOS and Mac OS X clients send the first TS they received in
Split-Include attributes as TSr, if we return 0.0.0.0/0 that either
results in a bunch of Quick Mode exchanges (for each TS) or an error
notify (ATTRIBUTES_NOT_SUPPORTED), depending on the client software and
version.  For Shrew and the Cisco client we still need to set the TS to
0.0.0.0/0 as they otherwise won't accept the Quick Mode response.

We basically have to return whatever TSr we received, instead of storing
them though, we simply make note of the IKE_SAs on which we received 0.0.0.0/0
as TSr.